+++ R/autospc_chart.R
@@ -84,6 +84,7 @@ autospc_chart_elements <- function() {
     "data",
     "x",
     "y",
+    "aggregation_na_rm",
     "period_min",
     "baseline_length",
     "shift_rule_threshold",
@@ -91,6 +92,7 @@ autospc_chart_elements <- function() {
     "establish_every_shift",
     "no_regrets",
     "overhanging_reversions",
+    "na_ends_run",
     "max_exclusions",
     "mr_screen_max_loops",
     "centre_line_tolerance",
@@ -253,6 +255,7 @@ assemble_chart_list <- function(
   data,
   x,
   y,
+  aggregation_na_rm = autospc_default("aggregation_na_rm"),
   period_min = autospc_default("period_min"),
   baseline_length = autospc_default("baseline_length"),
   shift_rule_threshold = autospc_default("shift_rule_threshold"),
@@ -260,6 +263,7 @@ assemble_chart_list <- function(
   establish_every_shift = autospc_default("establish_every_shift"),
   no_regrets = autospc_default("no_regrets"),
   overhanging_reversions = autospc_default("overhanging_reversions"),
+  na_ends_run = autospc_default("na_ends_run"),
   max_exclusions = autospc_default("max_exclusions"),
   mr_screen_max_loops = autospc_default("mr_screen_max_loops"),
   centre_line_tolerance = autospc_default("centre_line_tolerance"),
@@ -279,6 +283,7 @@ assemble_chart_list <- function(
     data = data,
     x = x,
     y = y,
+    aggregation_na_rm = aggregation_na_rm,
     period_min = period_min,
     baseline_length = as.integer(baseline_length),
     shift_rule_threshold = shift_rule_threshold,
@@ -286,6 +291,7 @@ assemble_chart_list <- function(
     establish_every_shift = establish_every_shift,
     no_regrets = no_regrets,
     overhanging_reversions = overhanging_reversions,
+    na_ends_run = na_ends_run,
     max_exclusions = max_exclusions,
     mr_screen_max_loops = mr_screen_max_loops,
     centre_line_tolerance = centre_line_tolerance,
@@ -356,6 +362,13 @@ n_effective_points.autospc_chart <- function(chart,
     nrow()
 
   return(points)
+}
+
+
+#' @noRd
+observed_rows.autospc_chart <- function(chart,
+                                        data) {
+  return(!is.na(data$y))
 }